refactor(logic): 分離した型定義を専用ファイルに移動
