+++ package/main/src/types/logic/equal.ts
@@ -1,0 +1,2 @@
+// XとYが等しいかどうかを判定する型
+export type Equal<X, Y> = X extends Y ? true : false;
